backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.1
python-dotenv==1.0.0
pydantic[email]==2.4.2
cloudinary==1.36.0
groq==0.3.2
python-dateutil==2.8.2
typing-extensions==4.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
mongomock==4.1.2
pytest-cov==4.1.0
starlette==0.36.3 